add form super class to avoid code duplication
